+++ src/frontend/context/behandlingContext/useBehandlingApi.ts
@@ -40,7 +40,7 @@ const useBehandlingApi = (
 
                     if (behandling.årsak === BehandlingÅrsak.SØKNAD) {
                         navigate(
-                            behandling.steg === BehandlingSteg.REGISTRERE_INSTITUSJON_OG_VERGE
+                            behandling.steg === BehandlingSteg.REGISTRERE_INSTITUSJON
                                 ? `/fagsak/${fagsakId}/${behandling?.behandlingId}/registrer-mottaker`
                                 : `/fagsak/${fagsakId}/${behandling?.behandlingId}/registrer-soknad`
                         );
