+++ src/frontend/komponenter/Fagsak/Vilkårsvurdering/Skjema/VilkårsvurderingSkjemaInstitusjon.tsx
@@ -17,7 +17,7 @@ import {
 } from '../../../../typer/vilkår';
 import PersonInformasjon from '../../../Felleskomponenter/PersonInformasjon/PersonInformasjon';
 import SamhandlerInformasjon from '../../../Felleskomponenter/SamhandlerInformasjon/SamhandlerInformasjon';
-import { useSamhandlerRequest } from '../../InstitusjonOgVerge/useSamhandler';
+import { useSamhandlerRequest } from '../../Institusjon/useSamhandler';
 import GeneriskAnnenVurdering from '../GeneriskAnnenVurdering/GeneriskAnnenVurdering';
 import GeneriskVilkår from '../GeneriskVilkår/GeneriskVilkår';
 import Registeropplysninger from '../Registeropplysninger/Registeropplysninger';
