+++ src/frontend/typer/fagsak.ts
@@ -1,7 +1,7 @@
 import type { BehandlingÅrsak } from './behandling';
 import type { BehandlingKategori, BehandlingUnderkategori } from './behandlingstema';
 import type { INøkkelPar } from './common';
-import type { IInstitusjon } from './institusjon-og-verge';
+import type { IInstitusjon } from './institusjon';
 import type { ITilbakekrevingsbehandling } from './tilbakekrevingsbehandling';
 import type { Utbetalingsperiode } from './vedtaksperiode';
 import type { VisningBehandling } from '../komponenter/Fagsak/Saksoversikt/visningBehandling';
