--- src/frontend/typer/institusjon-og-verge.ts
@@ -1,14 +1,0 @@
-export interface IVerge {
-    ident: string;
-}
-
-export interface IInstitusjon {
-    navn?: string;
-    orgNummer: string;
-    tssEksternId: string;
-}
-
-export interface IRegistrerInstitusjonOgVerge {
-    vergeInfo?: IVerge;
-    institusjonInfo?: IInstitusjon;
-}
