+++ src/frontend/typer/institusjon.ts
@@ -1,0 +1,10 @@
+export interface IInstitusjon {
+    navn?: string;
+    orgNummer: string;
+    tssEksternId: string;
+}
+
+export interface IRegistrerInstitusjon {
+    orgNummer: string;
+    tssEksternId: string;
+}
